+++ odoo/local-src/connector_esb/models/__init__.py
@@ -27,3 +27,4 @@ from . import partner_alcyon_category
 from . import sale_order
 from . import product_supplierinfo
 from . import queue_job
+from . import buyx_gety